+++ packages/api2/src/activate/activate.controller.ts
@@ -9,7 +9,7 @@ import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { ActivateInputDto } from './dto/activate-input.dto';
 import { User } from '@/users/schemas/users.schema';
 import { ActivateService } from './activate.service';
-import { MongooseClassSerializerInterceptor } from '@/shared/mongoose-class-serializer.interceptor';
+import { MongooseClassSerializerInterceptor } from '@/shared/interceptors/mongoose-class-serializer.interceptor';
 
 @Controller('activate')
 @ApiTags('Activate')
