+++ packages/api2/src/periodsettings/periodsettings.service.ts
@@ -1,4 +1,4 @@
-import { ServiceException } from '@/shared/service-exception';
+import { ServiceException } from '@/shared/exceptions/service-exception';
 import { UtilsProvider } from '@/utils/utils.provider';
 import { Inject, Injectable, forwardRef } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
