+++ packages/api2/src/quantifications/schemas/quantifications.schema.ts
@@ -3,7 +3,7 @@ import { Types } from 'mongoose';
 import { Praise } from '@/praise/schemas/praise.schema';
 import { User } from '@/users/schemas/users.schema';
 import { ApiResponseProperty } from '@nestjs/swagger';
-import { ExposeId } from '@/shared/expose-id.decorator';
+import { ExposeId } from '@/shared/decorators/expose-id.decorator';
 
 export type QuantificationDocument = Quantification & Document;
