+++ packages/api2/src/shared/exceptions/service-exception.ts
