+++ packages/api2/src/shared/interceptors/mongoose-class-serializer.interceptor.ts
