+++ packages/api2/test/activate.e2e-spec.ts
@@ -8,7 +8,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { AppModule } from '../src/app.module';
 import { Server } from 'http';
 import { Wallet } from 'ethers';
-import { ServiceExceptionFilter } from '@/shared/service-exception.filter';
+import { ServiceExceptionFilter } from '@/shared/filters/service-exception.filter';
 import { UsersSeeder } from '@/database/seeder/users.seeder';
 import { runDbMigrations } from '@/database/migrations';
 import { UsersModule } from '@/users/users.module';
